requirements.txt
# Dependencias del proyecto GIS Web App
# Instalar con: pip install -r requirements.txt

# Framework web
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Conexión a base de datos SQL Server
pyodbc>=4.0.39
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.0.0

# Librerías básicas para GIS
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.14.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0

# Librerías para geocodificación
requests>=2.31.0

# Librerías adicionales útiles para GIS
rasterio>=1.3.0
fiona>=1.9.0
pyproj>=3.6.0
contextily>=1.4.0

# Servidor WSGI para producción (Windows/IIS)
waitress>=2.1.2

# Librerías para desarrollo
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
